Edit composer.json
