+++ composer.json
@@ -7,7 +7,7 @@
     "minimum-stability": "dev",
     "require": {
         "php": "^5.4|^7.0",
-        "vkcom/vk-api-schema": "1.4.1"
+        "vkcom/vk-api-schema": "@dev"
     },
     "require-dev": {
       "phpunit/phpunit": "~4.0"
